+++ transcendence-app/src/app.module.ts
@@ -5,7 +5,7 @@ import { validate } from './config/env.validation';
 import { APP_INTERCEPTOR } from '@nestjs/core';
 import { SocketModule } from './socket/socket.module';
 import { ChatModule } from './chat/chat.module';
-import { AuthorizationModule } from './shared/authorization/authorization.module';
+import { AuthorizationModule } from './authorization/authorization.module';
 
 @Module({
   imports: [
