+++ transcendence-app/src/authorization/infrastructure/authuser.chatroom-member.ts
