+++ transcendence-app/src/authorization/infrastructure/authuser.chatroom.ts
