+++ transcendence-app/src/authorization/infrastructure/authuser.global.ts
