+++ transcendence-app/src/shared/db/db.module.ts
@@ -18,8 +18,8 @@ import { IChatMessageRepository } from '../../chat/chat/infrastructure/db/chat-m
 import { ChatMessagePostgresRepository } from '../../chat/chat/infrastructure/db/postgres/chat-message.postgres.repository';
 import { IFriendRepository } from '../../user/infrastructure/db/friend.repository';
 import { FriendPostgresRepository } from '../../user/infrastructure/db/postgres/friend.postgres.repository';
-import { IUserToRoleRepository } from '../../user/infrastructure/db/user-to-role.repository';
-import { UserToRolePostgresRepository } from '../../user/infrastructure/db/postgres/user-to-role.postgres.repository';
+import { IUserToRoleRepository } from '../../authorization/infrastructure/db/user-to-role.repository';
+import { UserToRolePostgresRepository } from '../../authorization/infrastructure/db/postgres/user-to-role.postgres.repository';
 
 @Module({
   providers: [
